+++ packages/moonstone/Marquee/MarqueeDecorator.js
@@ -1,3 +1,4 @@
+import deprecate from '@enact/core/internal/deprecate';
 import hoc from '@enact/core/hoc';
 import {forward} from '@enact/core/handle';
 import {childrenEquals} from '@enact/core/util';
@@ -110,6 +111,14 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 
 		static propTypes = /** @lends moonstone/Marquee.MarqueeDecorator.prototype */ {
 			/**
+			 * Text alignment value of the marquee. Valid values are `'left'`, `'right'` and `'center'`.
+			 *
+			 * @type {String}
+			 * @public
+			 */
+			alignment: PropTypes.oneOf(['left', 'right', 'center']),
+
+			/**
 			 * Children to be marqueed
 			 *
 			 * @type {Node}
@@ -141,6 +150,7 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			 *
 			 * @type {Boolean}
 			 * @public
+			 * @deprecated replaced by `alignment`
 			 */
 			marqueeCentered: PropTypes.bool,
 
@@ -229,6 +239,10 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			this.forceRestartMarquee = false;
 
 			this.invalidateMetrics();
+
+			if (this.props.marqueeCentered) {
+				deprecate({name: 'marqueeCentered', since: '1.7.0', message: 'Use `alignment` instead', until: '2.0.0'});
+			}
 		}
 
 		componentDidMount () {
@@ -529,10 +543,10 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 
 		renderMarquee () {
 			const {
+				alignment,
 				children,
 				disabled,
 				forceDirection,
-				marqueeCentered,
 				marqueeOn,
 				marqueeSpeed,
 				...rest
@@ -557,6 +571,7 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 				rest[enter] = this.handleEnter;
 			}
 
+			delete rest.marqueeCentered;
 			delete rest.marqueeDelay;
 			delete rest.marqueeDisabled;
 			delete rest.marqueeOnRenderDelay;
@@ -567,8 +582,8 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			return (
 				<Wrapped {...rest} disabled={disabled}>
 					<Marquee
+						alignment={alignment}
 						animating={this.state.animating}
-						centered={marqueeCentered}
 						className={marqueeClassName}
 						clientRef={this.cacheNode}
 						distance={this.distance}
@@ -587,6 +602,7 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 		renderWrapped () {
 			const props = Object.assign({}, this.props);
 
+			delete props.alignment;
 			delete props.marqueeCentered;
 			delete props.marqueeDelay;
 			delete props.marqueeDisabled;